\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity of resampling to estimated sequestration rates}
\description{Estimates sensitivity of resampling to estimated sequestration rates. }
\usage{sensitivity(x, ...)}
\arguments{
  \item{x}{object of type ??}
  \item{...}{optional argument}
}
\author{Ichsani Wheeler}
\seealso{
\code{\link{returns}}
}