\name{"SpatialStratifiedSample"-class}
\docType{class}
\alias{"SpatialStratifiedSample"-class}
\title{A class for stratified point samples}
\description{A class containing input and output maps generated by running the \code{stratification::strata.LH} method on a raster map.}
\section{Slots}{
  \describe{
    \item{\code{variable}:}{object of class \code{"character"}; variable name}
    \item{\code{locations}:}{object of class \code{"SpatialPoints"} optimized locations }
    \item{\code{strata}:}{object of class \code{"SpatialPixelsDataFrame"}; contains the output stratification and probabilities}
    \item{\code{LH}:}{object of class \code{"data.frame"}; contains the outputs of analysis produced by the stratification package}
    \item{\code{eval}:}{object of class \code{"data.frame"}; contains a list of total variances}
  }
}
\section{Methods}{
  \describe{
	\item{plot}{\code{signature(x = "SpatialStratifiedSample")}: plots map showing the samples and the optimisation curve}
  }
}
\author{ Ichsani Wheeler }
\seealso{ \code{\link{eval.LH}}, \code{stratification::strata.LH}
}
\keyword{classes}
